Custom hook useMap
