+++ project/src/components/map/map.tsx
@@ -1,7 +1,6 @@
 import ICity from '../../interfaces/ICity';
-import { useEffect, useRef, useState } from 'react';
-import leaflet from 'leaflet';
-// import useMap from '../../hooks/useMap';
+import { useRef } from 'react';
+import useMap from '../../hooks/useMap';
 
 type MapProps = {
   city: ICity
@@ -10,26 +9,7 @@ type MapProps = {
 function Map({city}: MapProps): JSX.Element {
   const mapRef = useRef(null);
 
-  const [map, setMap] = useState(null);
-
-  useEffect(() => {
-    if (mapRef.current) {
-      const mapInstance = leaflet.map(mapRef.current, {
-        center: {
-          lat: city.latitude,
-          lng: city.longitude,
-        },
-        zoom: city.zoom,
-      });
-
-      leaflet.tileLayer(
-        'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png', {
-          attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
-        }).addTo(mapInstance);
-
-      setMap(map);
-    }
-  }, [mapRef, city, map]);
+  const map = useMap(mapRef, city);
 
   return (
     <section ref={mapRef} className="cities__map map"></section>
