+++ project/src/hooks/useMap.js
@@ -1,0 +1,27 @@
+import { useState, useEffect } from 'react';
+import leaflet from 'leaflet';
+
+function useMap(mapRef, city) {
+  const [map, setMap] = useState(null);
+
+  useEffect(() => {
+    if (mapRef.current) {
+      const mapInstance = leaflet.map(mapRef.current, {
+        center: {
+          lat: city.latitude,
+          lng: city.longitude,
+        },
+        zoom: city.zoom,
+      });
+
+      leaflet.tileLayer(
+        'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png', {
+          attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
+        }).addTo(mapInstance);
+
+      setMap(map);
+    }
+  }, [mapRef, city, map]);
+}
+
+export default useMap;
